requirements.txt
loguru
torch
pyyaml
numpy
